feat: update clock to expose epoch number in event
